move 'equals' tests to separate file
